+++ src/test/java/com/pax/market/api/sdk/java/api/terminal/ResellerApiTest.java
@@ -12,19 +12,17 @@
 package com.pax.market.api.sdk.java.api.terminal;
 
 import java.util.LinkedHashMap;
-import java.util.Locale;
-
+import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import com.pax.market.api.sdk.java.api.base.dto.Result;
 import com.pax.market.api.sdk.java.api.reseller.ResellerApi;
 import com.pax.market.api.sdk.java.api.reseller.ResellerApi.ResellerSearchOrderBy;
-import com.pax.market.api.sdk.java.api.reseller.ResellerApi.ResellerStatus;
 import com.pax.market.api.sdk.java.api.reseller.dto.ResellerCreateRequest;
 import com.pax.market.api.sdk.java.api.reseller.dto.ResellerDTO;
 import com.pax.market.api.sdk.java.api.reseller.dto.ResellerPageDTO;
 import com.pax.market.api.sdk.java.api.reseller.dto.ResellerUpdateRequest;
+import com.pax.market.api.sdk.java.api.test.TestConstants;
 import com.pax.market.api.sdk.java.api.util.EnhancedJsonUtils;
 
 /**
@@ -39,12 +37,12 @@ public class ResellerApiTest {
 	
 	static Long newResellerId = 0L;
 	
-    public static void init(){
-//    	resellerApi = new  ResellerApi("https://api.whatspos.cn/p-market-api/", "ZJFXJAG7SJXPPESKVAPO", "AXN5ES2BFYYY8FRMSAPXKQ2ZMT22WYTQGCOGGFM9");
-    	resellerApi = new  ResellerApi("http://localhost:8080/p-market-api/", "WZ5JT4WYOCCN0JQ7UZEV", "ALRBEIJDE9FDJHPWGCSGOTCM57HBDQFU0U2VGQPP");
+    public void init(){
+    	resellerApi = new  ResellerApi(TestConstants.API_BASE_URL, TestConstants.API_KEY, TestConstants.API_SECRET);
     }
     
-    public static void testSearchReseller() {
+    @Test
+    public void testSearchReseller() {
     	Result<ResellerPageDTO> result = resellerApi.searchReseller(1, 10, ResellerSearchOrderBy.Contact, null, null);
     	logger.info(result.toString());
         
@@ -52,15 +50,17 @@ public class ResellerApiTest {
     	System.out.println("search result="+result);
     }
     
-    public static void testGetReseller() {
+    @Test
+    public void testGetReseller() {
     	Result<ResellerDTO> result = resellerApi.getReseller(1000005210L);
     	System.out.println(EnhancedJsonUtils.toJson(result));
     	System.out.println("get result="+result);
     }
     
-    public static void testCreateReseller() {
+    @Test
+    public void testCreateReseller() {
     	ResellerCreateRequest request = new ResellerCreateRequest();
-    	request.setName("reseller_001");
+    	request.setName("reseller_003");
     	request.setContact("Sam");
     	request.setCountry("CN");
     	request.setEmail("sam2@gmail.com");
@@ -68,7 +68,8 @@ public class ResellerApiTest {
     	request.setPostcode("850212");
     	request.setCompany("Cam");
     	request.setAddress("JiangSu Suzhou city xinghujie 203#");
-    	request.setParentResellerName("Jesse");
+    	request.setParentResellerName("reseller");
+    	request.setActivateWhenCreate(Boolean.FALSE);
 //    	LinkedHashMap<String,String> attrs = new LinkedHashMap<String,String>();
 //    	attrs.put("111", "tan2");
 //    	request.setEntityAttributeValues(attrs);
@@ -80,7 +81,8 @@ public class ResellerApiTest {
     	System.out.println("create reseller result="+result);
     }
     
-    public static void testUpdateReseller() {
+    @Test
+    public void testUpdateReseller() {
     	ResellerUpdateRequest request = new ResellerUpdateRequest();
     	request.setName("FVFFF22fef");
     	request.setContact("FFF");
@@ -106,43 +108,32 @@ public class ResellerApiTest {
     	System.out.println(EnhancedJsonUtils.toJson(result));
     }
     
-    public static void testActivateReseller() {
+    @Test
+    public void testActivateReseller() {
     	Result<String> result = resellerApi.activateReseller(1000005211L);
     	System.out.println("activate result="+result);
     	System.out.println(EnhancedJsonUtils.toJson(result));
     }
     
-    public static void testDisableReseller() {
+    @Test
+    public void testDisableReseller() {
     	Result<String> result = resellerApi.disableReseller(1000005211L);
     	System.out.println("disable result="+result);
     	System.out.println(EnhancedJsonUtils.toJson(result));
     }
     
-    public static void testDeleteReseller() {
+    @Test
+    public void testDeleteReseller() {
     	Result<String> result = resellerApi.deleteReseller(1000005212L);
     	System.out.println("Delete result="+result);
     	System.out.println(EnhancedJsonUtils.toJson(result));
     }
     
-    public static void testReplaceEmail() {
+    @Test
+    public void testReplaceEmail() {
     	Result<String> result = resellerApi.replaceResellerEmail(1000005212L, "zhangsan@pax.com");
     	System.out.println("Replace email result="+result);
     	System.out.println(EnhancedJsonUtils.toJson(result));
     }
     
-    public static void main(String[] args){
-        init();
-//        testSearchReseller();
-//        
-//        testCreateReseller();
-//        testActivateReseller();
-        testUpdateReseller();
-//        testGetReseller();
-//        testActivateReseller();
-//        testDisableReseller();
-//        testDeleteReseller();
-        
-//        testReplaceEmail();
-        
-    }
 }