+++ src/test/java/com/pax/market/api/sdk/java/api/test/TestConstants.java
@@ -17,8 +17,14 @@ package com.pax.market.api.sdk.java.api.test;
  * @date  2018/11/02 15:23:22
  */
 public interface TestConstants {
-
-	public static final String API_BASE_URL = "https://api.staging.whatspos.com/p-market-api";
-	public static final String API_KEY = "MCXF60RUVJPKWGBGATYO";
-	public static final String API_SECRET = "OF93E3I7A612QHAA55A56HNQ908QYL3MPUTZK6WU";
+	
+	public static final String API_BASE_URL = "https://api.whatspos.com/p-market-api";
+	public static final String API_KEY = "NGUIWQGJZGMGSQD0SNA5";
+	public static final String API_SECRET = "WX905C6V8K2NGDC9RBGVPH7NIARQSGUA4VVV746O";
+	
+//	public static final String API_BASE_URL = "http://localhost:8080/p-market-api";
+//	public static final String API_KEY = "NGUIWQGJZGMGSQD0SNA5";
+//	public static final String API_SECRET = "WX905C6V8K2NGDC9RBGVPH7NIARQSGUA4VVV746O";
+	
+	
 }